requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
Django==5.1
django-allauth==64.1.0
django-crispy-forms==2.3
idna==3.7
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
